I've checked the last versions of the figure scripts in my computer and in the OwnCloud, and cleaned it up a bit the ones I think/hope are the last ones:
I manage to replicate the figures 1 and 2 perfectly, using the latest versions of the .Rda data files that Michael uploaded in OwnCloud  2-3 mo ago.
I know the codes for figure 4 and S1 work well too with the 'diversity' Rda that you produced in R. But I deleted this object from my computer because it was >1.5GB. I'v checked the code and it looks like it works.
I've checked the code for Fig 3 and *I think* it works fine, but we will need Michael here. He wrote this code and I've never run it. Michael, could you please check with your Rda files?

Pending:
fix the full paths: in my scripts, these are calling folders in my machine. Please see that in my figure codes, section 2 usually calls the datasets. We need to either deposit the Rda files somewhere  (in github or Zenodo or elsewhere) and standardize all codes to call those files --or explain that the code to generate this data is the point of the repo...
We (you?) could make the figure codes more efficient and prettier. I've tried to clean up my comments, but it's still clunky. I've seen that you created a neat code to produce the networks, while my code runs one by one again. It's functional, but not beautiful. I like functional things. If you feel like standardizing all codes, go for it