Comment by Reviewer: Figure S1, the different types of lines are hard to differentiate, maybe consider using different colours for the four possible traits (instead of using it for the different network types, which are already clearly delineated by the graph plots and the headings).((
Our changes: We have replotted this figure to (i) set a simpler colour code with high contrast (red vs. black) to highlight the differences between the two cultural lineages (A and B), (ii) improve the distinction between the line thicknesses and types (dotted, dashed, dotted+dashed, full lines) to highlight the differences between the cultural traits (single item, dyad, triad, crossover), (iii) increase the readability of the plots by adding a label to the y-axis (the proportion of the population with a given cultural trait) and a legend at the bottom to point out clearly to all types of cultural traits.